refactor: split backend.rs and listen/mobile.rs under the 800-line cap (#1672)
